Modificaciones en los servicios y s eagregaron 2 vistas

+++ src/main/java/ar/edu/unlam/tallerweb1/dao/EstadoMovimientoDao.java
@@ -7,5 +7,5 @@ import ar.edu.unlam.tallerweb1.modelo.EstadoMovimiento;
 
 public interface EstadoMovimientoDao {
 	List<EstadoMovimiento> listarEstadosMovimientos();
-	EstadoMovimiento buscarPorId(Long id);
+	EstadoMovimiento buscarPorId(int id);
 }
